Merge the branch/blackhole-improvement (via the blackhole-improvement-merge branch).
The basic goal is to split pyjitpl.py in two: the
normal tracer, and a faster version (in blackhole.py)
dedicated to running in blackhole mode.  The latter
does no tracing, but is much faster.  Gives good
performance gains on examples which need to do a lot
of blackholing.

The number of changes is a bit huge though.  The
format of the static bytecodes used by the jit
changed completely, and codewriter.py is now split
among many files in the new directory pypy.jit.codewriter.
There is also no longer ConstAddr, only ConstInt: prebuilt
addresses are now turned into integers (which are
symbolic, with the new class AddressAsInt, so they can
be rendered by the C translation backend).  Various
related changes occurred here and there.


git-svn-id: http://codespeak.net/svn/pypy/trunk@75220 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada