issue545 testing
Change CALL_PURE into a temporary operation.  The idea is to fix issue545
by preventing multiple calls to a pure function.  Now a call to a pure
function is traced as a normal CALL, then immediately turned into a
CALL_PURE by whacking at the history, and the value that it returned
is inserted as first argument.  If optimizeopt can constant-fold it
away, it thus has the returned value available.  If it cannot, then it
will turn it again into a regular CALL.


git-svn-id: http://codespeak.net/svn/pypy/trunk@75229 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada