Merge branch/int-between.
It adds an 'int_between' operation produced by lltypesystem.rclass.
This operation has a slightly more efficient form in the C backend.
More importantly, in the JIT it becomes a single True/False test;
previously, it was done with a double test "a <= b <= c", which
can fail for two distinct reasons: b < a or b > c.  If we ever see
that case, it would create two equal traces from that point.


git-svn-id: http://codespeak.net/svn/pypy/trunk@75320 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada